quiz/requirements.txt
asgiref==3.6.0
certifi==2023.5.7
charset-normalizer==2.0.12
Django==3.2.16
djangorestframework==3.13.1
flake8==5.0.4
gunicorn==20.0.4
idna==3.4
importlib-metadata==4.2.0
mccabe==0.7.0
psycopg2-binary==2.8.6
pycodestyle==2.9.1
pyflakes==2.5.0
pytz==2023.3
requests==2.27.1
sqlparse==0.4.4
typing_extensions==4.6.0
urllib3==1.26.15
zipp==3.15.0
